Move PostListPagination to component folder
