+++ src/templates/blog-list.js
@@ -3,8 +3,7 @@ import Layout from "../components/layout"
 import { graphql, Link } from "gatsby"
 import style from "./blog-list.module.css"
 import { getTitle } from "../utils/general"
-import PostListPagination from "../components/post-list-pagination"
-import { AiFillTag } from "react-icons/all"
+import PostListPagination from "../components/PostListPagination/PostListPagination"
 
 export default ({ data, pageContext }) => {
   const posts = data.allMarkdownRemark.edges
