+++ src/templates/project-list.js
@@ -3,7 +3,7 @@ import Layout from "../components/layout"
 import {graphql, Link} from "gatsby"
 import projectStyles from "./project-list.module.css"
 import {getTitle} from "../utils/general"
-import PostListPagination from "../components/post-list-pagination"
+import PostListPagination from "../components/PostListPagination/PostListPagination"
 
 export default ({data, pageContext}) => {
   const posts = data.allProjectsJson.edges
